+++ src/singleTable/model/from/fromEntity/definitions/query/common.ts
@@ -1,9 +1,11 @@
 import type { QueryParams } from 'provider';
-import type { AnyObject } from 'types';
+import type { AnyObject, HasUndefined } from 'types';
 
 export type QueryConfigParams<Entity = AnyObject> = Omit<
   QueryParams<Entity>,
   'index' | 'partitionKey' | 'rangeKey' | 'table'
 >;
 
-export type OptionalTupleIfUndefined<Ref, Params> = Ref extends undefined ? [Params?] : [Params];
+export type OptionalTupleIfUndefined<Ref, Params> = HasUndefined<[Ref]> extends true
+  ? [Params?]
+  : [Params];